# BE/openapi.yaml
openapi: 3.0.3
info:
  title: JHIC2.0 CS Chatbot API
  version: 1.0.0
  description: Prototype backend chatbot ketat-data + mini-router. Source of truth kontrak FE-BE.
servers:
  - url: http://localhost:3000
tags:
  - name: chat
  - name: knowledge
  - name: health
paths:
  /api/v1/chat:
    post:
      tags: [chat]
      summary: Kirim pesan chat (alias simpel)
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema: { type: string }
          description: Retry aman. Key sama + body sama = replay. Key sama + body beda = 409.
        - in: header
          name: X-User-Id
          required: false
          schema: { type: string }
          description: Identitas user untuk rate limit. Fallback IP+session_id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, message]
              properties:
                session_id: { type: string, pattern: '^[A-Za-z0-9_-]+$', maxLength: 64, example: sess_demo01 }
                message: { type: string, minLength: 1, maxLength: 1000, example: Jam operasional CS? }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id: { type: string, example: msg_a1b2c3d4 }
                      session_id: { type: string }
                      reply: { type: string }
                      sources: { type: array, items: { type: object, properties: { id: { type: string }, score: { type: number } } } }
                      provider: { type: string, example: openrouter }
                      model: { type: string }
                      fallback: { type: boolean }
                      at: { type: string, format: date-time }
        '400': { description: Validasi gagal }
        '409': { description: Idempotency-Key dipakai ulang dengan body beda }
        '429':
          description: Rate limited (15/menit/user, 300/menit global)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code: { type: string, example: RATE_LIMITED }
                      message: { type: string }
                      retry_after: { type: integer, example: 23 }
                      scope: { type: string, enum: [user_min, user_hour, global_min] }
  /api/v1/conversations/{id}/messages:
    post:
      tags: [chat]
      summary: Kanonis resource-oriented (sama dengan /chat)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string }
      responses:
        '200': { description: OK }
  /api/v1/knowledge:
    get:
      tags: [knowledge]
      summary: List FAQ dengan cursor pagination
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 10, minimum: 1, maximum: 50 }
        - in: query
          name: cursor
          schema: { type: string, description: Opaque cursor dari next_cursor }
      responses:
        '200': { description: OK }
  /api/v1/health:
    get:
      tags: [health]
      summary: Status + faq_count + status router
      responses:
        '200': { description: OK }
components: {}